Fixed some misbehaving floats

+++ main.js
@@ -2,7 +2,7 @@
 
 var currency = require('y-currency');
 
-var amount = parseInt(process.argv[2]);
+var amount = parseFloat(process.argv[2]);
 var from = process.argv[3].toString();
 var to = process.argv[4].toString();
